do catastrophes
Make the AI perform catastrophes automatically whenever possible.
Eventually we would want to be strategic about this--for example, not
calling a catastrophe on your own pieces in the middle of a large
sacrifice action--but for now we'll just be pessimistic.

The AI now plays against itself pretty well (read: not obviously badly)
but eventually gets stuck in a capture loop after a few dozen moves,
after both players have built up their fleets.

Also fix a bug in Largest.