no takebacks, muck with score weights, add transcript
Don't allow a move to undo the previous player's move. This prevents the
capture loop that we were getting into. In practice, the only move that
can return the game to prior state is attackingo

Change the scoring algorithm to give even more weight to larger ships.
Not sure if this makes any appreciable difference.

Finally, add an interesting transcript of a game between AI players:
around move 39, north makes a move which has a surprising effect: it
causes a catastrophe in south's homeworld which destroys part of the
star and makes it connect directly to north's homeworld. When south
moves a red ship to north's homeworld on the next turn, south's victory
is assured... except the game goes on for another hundred moves because
south isn't smart enough to make the winning move immediately.