prefer earlier wins, switch to negamax
Rate earlier moves higher so that minimax will always pick
the winning move if possible. Before, it would merely pick any move
which did not forfeit the win; this would lead to a situation where,
even though the game was already decided, it would continue
indefinitely, with one of the players always being "a couple moves"
away from winning.