fix nondeterminism in the order homeworlds are printed
same problem as before: homewords are stored in a map, and the order of
a map isn't defined. i can't believe i fixed the other loop but missed
this one! in my defense, this bug cropped up way less often because
there are only so many ways you can reorder two things.