integrate sacrifice actions into minimax
this probably isn't as efficient as it could be, but it's a start.

note that i had to disable Y3 sacrifices because the search space was
too huge.

also: fix an accounting error when gc'ing a catastrophe at a homeworld.
the game is over at that point, so it shouldn't matter, but it was
causing false positives in the sanity check.